requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
discord.py==2.3.1
requests==2.31.0